refactor: move dialog windows to a separate folder
