+++ app/src/main/kotlin/Main.kt
@@ -1,5 +1,4 @@
 import androidx.compose.desktop.ui.tooling.preview.Preview
-import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.window.Window
@@ -7,7 +6,7 @@ import androidx.compose.ui.window.WindowPosition
 import androidx.compose.ui.window.application
 import androidx.compose.ui.window.rememberWindowState
 import java.awt.Dimension
-import view.utils.SelectInitDialogWindow
+import view.utils.dialogWindows.SelectInitDialogWindow
 
 @Composable
 @Preview
