+++ app/src/main/kotlin/view/tabScreen/FileControlTab.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.window.Dialog
 import kotlinx.coroutines.delay
 import model.io.neo4j.Neo4jRepositoryHandler
 import model.io.sql.SQLDatabaseModule
-import view.utils.*
+import view.utils.dialogWindows.*
 import java.awt.FileDialog
 import java.awt.Frame
