+++ app/src/main/kotlin/view/utils/dialogWindows/SelectInitDialogWindow.kt
@@ -1,4 +1,4 @@
-package view.utils
+package view.utils.dialogWindows
 
 import MyAppTheme
 import androidx.compose.foundation.background
@@ -15,6 +15,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
+import view.utils.dialogWindows.CreateGraphDialogWindow
+import view.utils.dialogWindows.ImportGraphDialogWindow
 import viewmodel.graph.SetupGraphViewModel
